+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [3.0.8] - 2025-05-06
 ### Features
-- Add CTRL/ALT modifier on Initiative button (only with MidiQoL). (Related Issue: [#158])
+- Add CTRL/ALT modifier on Initiative button. (Related Issue: [#158])
 
 ## [3.0.7] - 2025-05-06
 ### Fixed